+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 1.0.1
 
-* This is the first version of the package. This will show a basic popup whenever you publish a new update of your app on Play Store.
+- This is the first version of the package. This will show a basic popup whenever you publish a new update of your app on Play Store.
+
+## 1.0.2
+
+- Feature added to get update details so that users can create their own widget to notify their users about a new update.